requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-extensions==2.2.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
idna==2.8
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
